postgres_to_es/requirements.txt
annotated-types==0.6.0
anyio==4.3.0
astroid==3.1.0
async-timeout==4.0.3
backports.zoneinfo==0.2.1
certifi==2024.2.2
charset-normalizer==3.3.2
dill==0.3.8
exceptiongroup==1.2.1
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
isort==5.13.2
load-dotenv==0.1.0
mccabe==0.7.0
pkg_resources==0.0.0
platformdirs==4.2.0
psycopg==3.1.18
pydantic==2.7.0
pydantic_core==2.18.1
pylint==3.1.0
python-dotenv==1.0.1
redis==5.0.3
sniffio==1.3.1
tomli==2.0.1
tomlkit==0.12.4
typing_extensions==4.11.0
urllib3==2.2.1
